requirements/base.txt
pytz==2021.3  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==9.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.3.0  # https://github.com/evansd/whitenoise
redis==4.1.2  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.3  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.17.0.post1  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.2.11  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.47.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.21.1

# HTML Minify and Tailwind CSS
django-htmlmin==0.11.0 # https://django-tailwind.readthedocs.io/en/latest/installation.html
django-tailwind==3.1.1 # https://pypi.org/project/django-tailwind/

# WYSIWYG Editor
django-tinymce4-lite==1.8.0 # https://pypi.org/project/django-tinymce4-lite/
django-filebrowser-no-grappelli==4.0.0 #https://github.com/smacker/django-filebrowser-no-grappelli

# Admin Security with honeypot
django-admin-honeypot==1.1.0 # https://pypi.org/project/django-admin-honeypot/
django-jet2==1.0.12
FeedParser==5.2.1
google-api-python-client==1.4.1

django-stdimage==5.3.0 # http://johannes.maron.family/django-stdimage/

django-countries-plus==2.0.0

geocoder==1.38.1
ratelim==0.1.6

# Language packages for django
django-parler==2.3
django-rosetta==0.9.8
loguru==0.5.3

SQLAlchemy==1.4.27

django-social-share==2.2.1

django-mathfilters==1.0.0

# Location picker with maximindb
geoip2==4.5.0
django-user-agents==0.4.0
ua-parser==0.10.0
user-agents==2.2.0

# show map in django admin
django-leaflet==0.28.2

# reactlike components and zero reload for django
django-unicorn==0.42.1

django-autoslug==1.9.8
pydub==0.25.1
